+++ CHANGELOG.rst
@@ -1,9 +1,12 @@
 fish ?.?.? (released ???)
 =========================
 
+Interactive improvements
+------------------------
+- Builtin and function commands can now be colored separately via new variables :envvar:`fish_color_builtin` and :envvar:`fish_color_function` (:issue:`12837`).
+
 Regression fixes:
 -----------------
-
 - ``abbr --position=anywhere`` completions are no longer offered in argument position, to avoid cluttering the completion pager (:issue:`12838`).
 
 fish 4.8.0 (released June 24, 2026)